Included solution file, and d.suo. Finished with probabaility calualtion.

+++ Basic/Constants.cs
@@ -3,6 +3,6 @@
     public class Constants
     {
         // Initial value of pheromone on graph. C
-        public static readonly decimal MinimalValueOfPheromone = 0.00001M;
+        public static readonly double MinimalValueOfPheromone = 0.00001D;
     }
 }
